python/life-tracker/agent-onboarding/requirements.txt
# Dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.10.0  # Atualizado para versão compatível
python-dotenv==1.0.0

# Agno Framework
agno
# OpenAI
openai==1.60.0
# Anthropic
# anthropic==0.9.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1
tavily-python
groq

# Banco de dados
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
aiosqlite==0.21.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
mongoengine==0.28.1
pymongo==4.10.1


# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilitários
python-multipart==0.0.6
pydantic-settings==2.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Logging e monitoramento
structlog==23.2.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0


# Desenvolvimento
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
